multipathd: more robust command parsing
The current parser allows commands like "path sda show" or "list list
add path sda", because it doesn't enforce ordering of the keywords
and simply adds rather than or-s the values of the keys.

Rework this by using a separate byte for the "verb" and the allowed
qualifiers in the different positions of the command. No command
needs more than 3 qualifiers, so we disallow command vectors with
more than 4 elements, and change the type of the fingerprint to
uint32_t.

init_callbacks() is now required in both multipathc and multipathd
to initialize the table of handlers and fingerprints. But as m
ultipathc never calls any handler, we need not link the
entire handlers code into it.

+++ multipathd/multipathc.c
@@ -40,7 +40,7 @@
 
 #if defined(USE_LIBREADLINE) || defined(USE_LIBEDIT)
 static int
-key_match_fingerprint (struct key * kw, uint64_t fp)
+key_match_fingerprint (struct key * kw, uint32_t fp)
 {
 	if (!fp)
 		return 0;
@@ -55,7 +55,7 @@ char *
 key_generator (const char * str, int state)
 {
 	static int index, len, has_param;
-	static uint64_t rlfp;
+	static uint32_t rlfp;
 	struct key * kw;
 	int i;
 	struct handler *h;
@@ -127,7 +127,7 @@ key_generator (const char * str, int state)
 			 * nfp is the candidate fingerprint we try to
 			 * validate against all known command fingerprints.
 			 */
-			uint64_t nfp = rlfp | kw->code;
+			uint32_t nfp = rlfp | kw->code;
 			vector_foreach_slot(handlers, h, i) {
 				if (!rlfp || ((h->fingerprint & nfp) == nfp)) {
 					/*
